d8

Description. (from https://groupprops.subwiki.org/wiki/Dihedral_group:D8#Geometric_definition)

Usage

d8

Format

An array with 8 rows and 4 variables:

w	x	y	z	op	cycles	description

1	2	3	4	e	()	noop

2	3	4	1	a	(1234)	rotation by 90 counter-clockwise

3	4	1	2	a2	(13)(24)	rotation by 180

4	1	2	3	a3	(1432)	rotation by 270 counter-clockwise

3	2	1	4	x	(13)	reflection about the diagonal 2-4

4	3	2	1	ax	(14)(23)	reflection about the line joining midpoints of sides 14 and 23

1	4	3	2	a2x	(24)	reflection about the diagonal 1-3

2	1	4	3	a3x	(12)(34)	reflection about the line joining midpoints of sides 12 and 34

Source

SageMath 9.5:

G = DihedralGroup(4) print(gap.StructureDescription(G), gap.IdGroup(G)) for per in G: d = per.dict() print(tuple(d.values()))

Details

The dihedral group D_8 (also called D_4) is defined as the group of all symmetries of the square (the regular 4-gon).

This has a cyclic subgroup comprising rotations (which is the cyclic subgroup generated by a) and has four reflections each being an involution: reflections about lines joining midpoints of opposite sides, and reflections about diagonals.

Examples

e <- layoutsquares::d8[1,]
a  <- layoutsquares::d8[2,]
a2 <- layoutsquares::d8[3,]
a3 <- layoutsquares::d8[4,]
x <- layoutsquares::d8[5,]
ax <- layoutsquares::d8[6,]
a2x <- layoutsquares::d8[7,]
a3x <- layoutsquares::d8[8,]

d8ops <- list(e, a, a2, a3, x, ax, a2x, a3x)

# https://groupprops.subwiki.org/wiki/Dihedral_group:D8#Multiplication_table
multiplication_table <- list(
  list(e, a, a2, a3, x, ax, a2x, a3x),
  list(a, a2, a3, e, ax, a2x, a3x, x),
  list(a2, a3, e, a, a2x, a3x, x, ax),
  list(a3, e, a, a2, a3x, x, ax, a2x),
  list(x, a3x, a2x, ax, e, a3, a2, a),
  list(ax, x, a3x, a2x, a, e, a3, a2),
  list(a2x, ax, x, a3x, a2, a, e, a3),
  list(a3x, a2x, ax, x, a3, a2, a, e)
)

for(i in 1:8) {
  for(j in 1:8) {
    opi <- d8ops[[i]]
    opj <- d8ops[[j]]
    prod <- multiplication_table[[i]][[j]]
    print(sprintf("%d,%d: c(%s)[c(%s)] ==  c(%s)", i,j,
      stringr::str_c(opi, collapse=','),
      stringr::str_c(opj, collapse=','),
      stringr::str_c(prod, collapse=',')))
    stopifnot(all(opi[opj] == prod))
  }
}
#> [1] "1,1: c(1,2,3,4)[c(1,2,3,4)] ==  c(1,2,3,4)"
#> [1] "1,2: c(1,2,3,4)[c(2,3,4,1)] ==  c(2,3,4,1)"
#> [1] "1,3: c(1,2,3,4)[c(3,4,1,2)] ==  c(3,4,1,2)"
#> [1] "1,4: c(1,2,3,4)[c(4,1,2,3)] ==  c(4,1,2,3)"
#> [1] "1,5: c(1,2,3,4)[c(3,2,1,4)] ==  c(3,2,1,4)"
#> [1] "1,6: c(1,2,3,4)[c(4,3,2,1)] ==  c(4,3,2,1)"
#> [1] "1,7: c(1,2,3,4)[c(1,4,3,2)] ==  c(1,4,3,2)"
#> [1] "1,8: c(1,2,3,4)[c(2,1,4,3)] ==  c(2,1,4,3)"
#> [1] "2,1: c(2,3,4,1)[c(1,2,3,4)] ==  c(2,3,4,1)"
#> [1] "2,2: c(2,3,4,1)[c(2,3,4,1)] ==  c(3,4,1,2)"
#> [1] "2,3: c(2,3,4,1)[c(3,4,1,2)] ==  c(4,1,2,3)"
#> [1] "2,4: c(2,3,4,1)[c(4,1,2,3)] ==  c(1,2,3,4)"
#> [1] "2,5: c(2,3,4,1)[c(3,2,1,4)] ==  c(4,3,2,1)"
#> [1] "2,6: c(2,3,4,1)[c(4,3,2,1)] ==  c(1,4,3,2)"
#> [1] "2,7: c(2,3,4,1)[c(1,4,3,2)] ==  c(2,1,4,3)"
#> [1] "2,8: c(2,3,4,1)[c(2,1,4,3)] ==  c(3,2,1,4)"
#> [1] "3,1: c(3,4,1,2)[c(1,2,3,4)] ==  c(3,4,1,2)"
#> [1] "3,2: c(3,4,1,2)[c(2,3,4,1)] ==  c(4,1,2,3)"
#> [1] "3,3: c(3,4,1,2)[c(3,4,1,2)] ==  c(1,2,3,4)"
#> [1] "3,4: c(3,4,1,2)[c(4,1,2,3)] ==  c(2,3,4,1)"
#> [1] "3,5: c(3,4,1,2)[c(3,2,1,4)] ==  c(1,4,3,2)"
#> [1] "3,6: c(3,4,1,2)[c(4,3,2,1)] ==  c(2,1,4,3)"
#> [1] "3,7: c(3,4,1,2)[c(1,4,3,2)] ==  c(3,2,1,4)"
#> [1] "3,8: c(3,4,1,2)[c(2,1,4,3)] ==  c(4,3,2,1)"
#> [1] "4,1: c(4,1,2,3)[c(1,2,3,4)] ==  c(4,1,2,3)"
#> [1] "4,2: c(4,1,2,3)[c(2,3,4,1)] ==  c(1,2,3,4)"
#> [1] "4,3: c(4,1,2,3)[c(3,4,1,2)] ==  c(2,3,4,1)"
#> [1] "4,4: c(4,1,2,3)[c(4,1,2,3)] ==  c(3,4,1,2)"
#> [1] "4,5: c(4,1,2,3)[c(3,2,1,4)] ==  c(2,1,4,3)"
#> [1] "4,6: c(4,1,2,3)[c(4,3,2,1)] ==  c(3,2,1,4)"
#> [1] "4,7: c(4,1,2,3)[c(1,4,3,2)] ==  c(4,3,2,1)"
#> [1] "4,8: c(4,1,2,3)[c(2,1,4,3)] ==  c(1,4,3,2)"
#> [1] "5,1: c(3,2,1,4)[c(1,2,3,4)] ==  c(3,2,1,4)"
#> [1] "5,2: c(3,2,1,4)[c(2,3,4,1)] ==  c(2,1,4,3)"
#> [1] "5,3: c(3,2,1,4)[c(3,4,1,2)] ==  c(1,4,3,2)"
#> [1] "5,4: c(3,2,1,4)[c(4,1,2,3)] ==  c(4,3,2,1)"
#> [1] "5,5: c(3,2,1,4)[c(3,2,1,4)] ==  c(1,2,3,4)"
#> [1] "5,6: c(3,2,1,4)[c(4,3,2,1)] ==  c(4,1,2,3)"
#> [1] "5,7: c(3,2,1,4)[c(1,4,3,2)] ==  c(3,4,1,2)"
#> [1] "5,8: c(3,2,1,4)[c(2,1,4,3)] ==  c(2,3,4,1)"
#> [1] "6,1: c(4,3,2,1)[c(1,2,3,4)] ==  c(4,3,2,1)"
#> [1] "6,2: c(4,3,2,1)[c(2,3,4,1)] ==  c(3,2,1,4)"
#> [1] "6,3: c(4,3,2,1)[c(3,4,1,2)] ==  c(2,1,4,3)"
#> [1] "6,4: c(4,3,2,1)[c(4,1,2,3)] ==  c(1,4,3,2)"
#> [1] "6,5: c(4,3,2,1)[c(3,2,1,4)] ==  c(2,3,4,1)"
#> [1] "6,6: c(4,3,2,1)[c(4,3,2,1)] ==  c(1,2,3,4)"
#> [1] "6,7: c(4,3,2,1)[c(1,4,3,2)] ==  c(4,1,2,3)"
#> [1] "6,8: c(4,3,2,1)[c(2,1,4,3)] ==  c(3,4,1,2)"
#> [1] "7,1: c(1,4,3,2)[c(1,2,3,4)] ==  c(1,4,3,2)"
#> [1] "7,2: c(1,4,3,2)[c(2,3,4,1)] ==  c(4,3,2,1)"
#> [1] "7,3: c(1,4,3,2)[c(3,4,1,2)] ==  c(3,2,1,4)"
#> [1] "7,4: c(1,4,3,2)[c(4,1,2,3)] ==  c(2,1,4,3)"
#> [1] "7,5: c(1,4,3,2)[c(3,2,1,4)] ==  c(3,4,1,2)"
#> [1] "7,6: c(1,4,3,2)[c(4,3,2,1)] ==  c(2,3,4,1)"
#> [1] "7,7: c(1,4,3,2)[c(1,4,3,2)] ==  c(1,2,3,4)"
#> [1] "7,8: c(1,4,3,2)[c(2,1,4,3)] ==  c(4,1,2,3)"
#> [1] "8,1: c(2,1,4,3)[c(1,2,3,4)] ==  c(2,1,4,3)"
#> [1] "8,2: c(2,1,4,3)[c(2,3,4,1)] ==  c(1,4,3,2)"
#> [1] "8,3: c(2,1,4,3)[c(3,4,1,2)] ==  c(4,3,2,1)"
#> [1] "8,4: c(2,1,4,3)[c(4,1,2,3)] ==  c(3,2,1,4)"
#> [1] "8,5: c(2,1,4,3)[c(3,2,1,4)] ==  c(4,1,2,3)"
#> [1] "8,6: c(2,1,4,3)[c(4,3,2,1)] ==  c(3,4,1,2)"
#> [1] "8,7: c(2,1,4,3)[c(1,4,3,2)] ==  c(2,3,4,1)"
#> [1] "8,8: c(2,1,4,3)[c(2,1,4,3)] ==  c(1,2,3,4)"

rm(e, a, a2, a3, x, ax, a2x, a3x, d8ops)